refactor(Graph): Move data explorer code to data explorer manager
#1292
